QA_CHATBOT_WITH_OLLAMA_AND_OPENAI/requirements.txt
langchain-openai
langchain
python-dotenv
langchain_core
langchain_community
streamlit